feat(core): 实现MTP线程池注解化配置与动态管理
- 新增Mtp注解支持字段和方法注入方式的线程池配置
- 实现MtpExecutorBeanPostProcessor处理线程池依赖注入
- 实现MtpRegisterExecutorBeanPostProcessor注册线程池到管理器
- 添加MtpThreadPoolExecutorProxy代理类增强线程池功能
- 集成Spring Boot Starter实现自动配置和生命周期管理
- 扩展UI组件显示实例数量从6个增加到9个
- 优化配置中心注册逻辑和状态同步机制
- 添加优雅关闭线程池工具方法支持超时控制
- 调整数据库表结构将keepAliveSeconds改为BIGINT类型

+++ mtp-config-center/src/main/java/com/mtp/config/center/entity/ThreadPoolConfigEntity.java
@@ -32,7 +32,7 @@ public class ThreadPoolConfigEntity {
     private Integer queueCapacity;
 
     @TableField("keep_alive_seconds")
-    private Integer keepAliveSeconds;
+    private Long keepAliveSeconds;
 
     @TableField("rejected_policy")
     private String rejectedPolicy;
